requirements.txt
ultralytics
matplotlib
numpy
opencv-python
pandas
tensorflow
scikit-learn
streamlit
dropbox
python-dotenv  